feature HWD-011 Fixes package on the IT module.
