+++ integration-test/src/main/kotlin/pt/pak3nuh/hollywood/sandbox/PetClinicException.kt
@@ -1,0 +1,3 @@
+package pt.pak3nuh.hollywood.sandbox
+
+class PetClinicException(message: String) : Exception(message)
