+++ integration-test/src/main/kotlin/pt/pak3nuh/hollywood/sandbox/actor/ParameterValidationActor.kt
@@ -1,0 +1,11 @@
+package pt.pak3nuh.hollywood.sandbox.actor
+
+import pt.pak3nuh.hollywood.processor.Actor
+
+@Actor
+interface ParameterValidationActor {
+
+//    suspend fun cannotReferenceOtherActors(clinicActor: ClinicActor)
+
+//    suspend fun cannotReturnAnActor(): VetActor
+}
