+++ integration-test/src/main/kotlin/pt/pak3nuh/hollywood/sandbox/actor/Randoms.kt
@@ -1,0 +1,18 @@
+package pt.pak3nuh.hollywood.sandbox.actor
+
+import kotlinx.coroutines.delay
+import kotlin.random.Random
+
+
+object Randoms {
+    private val random = Random(0)
+
+    fun int(until: Int): Int = random.nextInt(until)
+
+    suspend fun delay(until: Int) {
+        delay(int(until).toLong())
+    }
+
+    fun bool(): Boolean = random.nextBoolean()
+}
+
