+++ integration-test/src/main/kotlin/pt/pak3nuh/hollywood/sandbox/clinic/PetStatus.kt
@@ -1,0 +1,5 @@
+package pt.pak3nuh.hollywood.sandbox.clinic
+
+enum class OwnerContactResult {
+    NO_ACTION, APPLY_TREATMENT
+}
