+++ integration-test/src/main/kotlin/pt/pak3nuh/hollywood/sandbox/clinic/Receipt.kt
@@ -1,0 +1,5 @@
+package pt.pak3nuh.hollywood.sandbox.clinic
+
+import pt.pak3nuh.hollywood.sandbox.pet.PetId
+
+data class Receipt(val petId: PetId, val total: Int)
