+++ integration-test/src/main/kotlin/pt/pak3nuh/hollywood/sandbox/vet/CheckType.kt
@@ -1,0 +1,5 @@
+package pt.pak3nuh.hollywood.sandbox.vet
+
+enum class CheckType {
+    MRI, CT_SCAN, BLOOD_PANEL
+}
