+++ integration-test/src/main/kotlin/pt/pak3nuh/hollywood/sandbox/vet/Vet.kt
@@ -1,0 +1,3 @@
+package pt.pak3nuh.hollywood.sandbox.vet
+
+data class Vet(val name: String)
