--- integration-test/src/main/kotlin/pt/pakenuh/hollywood/sandbox/PetClinicException.kt
@@ -1,3 +1,0 @@
-package pt.pakenuh.hollywood.sandbox
-
-class PetClinicException(message: String) : Exception(message)
