--- integration-test/src/main/kotlin/pt/pakenuh/hollywood/sandbox/actor/ParameterValidationActor.kt
@@ -1,11 +1,0 @@
-package pt.pakenuh.hollywood.sandbox.actor
-
-import pt.pak3nuh.hollywood.processor.Actor
-
-@Actor
-interface ParameterValidationActor {
-
-//    suspend fun cannotReferenceOtherActors(clinicActor: ClinicActor)
-
-//    suspend fun cannotReturnAnActor(): VetActor
-}
