--- integration-test/src/main/kotlin/pt/pakenuh/hollywood/sandbox/clinic/Exam.kt
@@ -1,13 +1,0 @@
-package pt.pakenuh.hollywood.sandbox.clinic
-
-enum class Exam(val cost: Int) {
-    X_RAY(10),
-    SOUND_SCAN(20),
-    BLOOD_PRESSURE(5),
-    ALLERGY_TEST(50),
-    KIDNEY_EXAM(100)
-}
-
-sealed class ExamResult
-object OkResult : ExamResult()
-object NokResult : ExamResult()
