--- integration-test/src/main/kotlin/pt/pakenuh/hollywood/sandbox/clinic/PetStatus.kt
@@ -1,5 +1,0 @@
-package pt.pakenuh.hollywood.sandbox.clinic
-
-enum class OwnerContactResult {
-    NO_ACTION, APPLY_TREATMENT
-}
