--- integration-test/src/main/kotlin/pt/pakenuh/hollywood/sandbox/clinic/Receipt.kt
@@ -1,5 +1,0 @@
-package pt.pakenuh.hollywood.sandbox.clinic
-
-import pt.pakenuh.hollywood.sandbox.pet.PetId
-
-data class Receipt(val petId: PetId, val total: Int)
