--- integration-test/src/main/kotlin/pt/pakenuh/hollywood/sandbox/pet/PetId.kt
@@ -1,9 +1,0 @@
-package pt.pakenuh.hollywood.sandbox.pet
-
-import kotlinx.serialization.ContextualSerialization
-import kotlinx.serialization.Serializable
-import pt.pakenuh.hollywood.sandbox.owner.OwnerId
-import java.time.LocalDate
-
-@Serializable
-data class PetId(val registryId: String, val name: String, val ownerId: OwnerId, @ContextualSerialization val birthDate: LocalDate)
