--- integration-test/src/main/kotlin/pt/pakenuh/hollywood/sandbox/vet/CheckType.kt
@@ -1,5 +1,0 @@
-package pt.pakenuh.hollywood.sandbox.vet
-
-enum class CheckType {
-    MRI, CT_SCAN, BLOOD_PANEL
-}
