--- integration-test/src/main/kotlin/pt/pakenuh/hollywood/sandbox/vet/Vet.kt
@@ -1,3 +1,0 @@
-package pt.pakenuh.hollywood.sandbox.vet
-
-data class Vet(val name: String)
